+++ lib/yard-yaml.rb
@@ -4,3 +4,6 @@
 # YARD tries requiring several patterns; providing `yard-yaml` ensures
 # it can be loaded regardless of whether YARD attempts `yard-yaml` or `yard/yaml`.
 require_relative "yard/yaml"
+
+Yard::Yaml::Plugin.activate(ARGV)
+Yard::Yaml::Plugin.install_at_exit(ARGV)